+++ src/lsp/messages.rs
@@ -201,6 +201,9 @@ pub enum LSPResponseType {
 
 impl LSPResponseType {
     pub fn parse(&self, value: Value) -> SerdeResult<LSPResponse> {
+        if value == Value::Null {
+            return Ok(LSPResponse::Empty);
+        }
         Ok(match self {
             Self::Completion(.., line, cursor) => match from_value::<CompletionResponse>(value)? {
                 CompletionResponse::Array(arr) => LSPResponse::Completion(arr, line.to_owned(), *cursor),
@@ -242,6 +245,7 @@ pub enum LSPResponse {
     TokensPartial { result: SemanticTokensRangeResult, max_lines: usize },
     Definition(GotoDefinitionResponse),
     Declaration(GotoDeclarationResponse),
+    Empty,
     Error(String),
 }