+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - reworked select handles
 - reworked rendering logic
 - fix: empty selects are ignored
+- fix: Empty response handle
 
 # Version 0.6.3
 - added support for non ascii in text_field